feat: [US-001] - Retire production ownership of test-only replay file wrappers
