+++ docs/development/deadcode-baseline.txt
@@ -1,12 +1,6 @@
 internal/testpath/artifact_contract.go:235:6: unreachable func: ArtifactContractEntryByPath
 internal/testpath/artifact_contract.go:245:6: unreachable func: MustArtifactContractEntry
 internal/testpath/artifact_contract.go:255:6: unreachable func: MustClassifiedArtifactPathFromCaller
-pkg/replay/event_stream_artifact.go:119:6: unreachable func: ArtifactFromEventStreamFile
-pkg/replay/event_stream_artifact.go:138:6: unreachable func: SaveArtifactFromEventStreamFile
-pkg/replay/event_stream_artifact.go:182:6: unreachable func: hydrateArtifactFromAdjacentFactory
-pkg/replay/event_stream_artifact.go:211:6: unreachable func: adjacentFactoryDir
-pkg/replay/event_stream_artifact.go:227:6: unreachable func: mergeGeneratedFactoryMissingRuntimeFields
-pkg/replay/event_stream_artifact.go:357:6: unreachable func: rewriteArtifactFactoryEvents
 pkg/testutil/artifact_contract.go:41:6: unreachable func: MustClassifiedArtifactPath
 pkg/testutil/replay_harness.go:36:6: unreachable func: WithReplayHarnessDir
 pkg/testutil/replay_harness.go:44:6: unreachable func: WithReplayHarnessServiceOptions
