+++ docs/processes/development-guide-relevant-files.md
@@ -63,7 +63,9 @@ This inventory records the checked-in files and directories that the maintainer 
 | `pkg/config/factory_config_mapping_test.go` | Generated factory-config boundary regression surface | Boundary rejection regressions should stay focused here by asserting `generatedFactoryBoundaryErrorPrefix` plus the full path-specific retirement guidance string for top-level, nested `definition`, and nested `definition.cron` aliases. |
 | `pkg/config/openapi_factory_test.go` | Generated OpenAPI JSON boundary regression surface | When cleanup work must preserve generated-boundary error strings or prove the generated `Factory` JSON shape itself, keep the proof here with `GeneratedFactoryFromOpenAPIJSON(...)` assertions over `generatedFactoryBoundaryErrorPrefix`, exact field paths, replacement guidance, and marshaled `factoryapi.Factory` workstation keys. |
 | `pkg/config/portable_bundled_files.go` | Supported portable bundled-file collection and materialization seam | Export/read-back portability changes should extend this seam so disk-backed `scripts/`, `docs/`, and allowlisted helper files stay mapped onto the canonical `factory/...` bundled-file contract for portable `factory/` layouts and named persisted factories without broadening generic runtime copy semantics; thin authored manifests should rehydrate existing supported entries from disk here before any broader default collection runs, and portable script entries that omit inline content still need load-time mode normalization because they bypass the write path that already calls `os.Chmod(...)`. |
+| `pkg/replay/event_artifact.go` | Replay run-started boundary normalization owner | Historical replay event streams may carry retired generated-factory workstation shapes even though current authored config rejects them; normalize only the replay run-started boundary here so checked-in artifacts remain replayable without relaxing general `pkg/config` generated-boundary validation. |
 | `pkg/replay/event_artifact_test.go` | Replay generated-payload merge regression surface | Direct worker and workstation replace-or-append assertions belong here when the contract under change is generated replay payload content and deterministic merge order, so the tests can seed `factoryapi.Factory` values without filesystem-heavy runtime-config fixtures. |
+| `tests/functional/replay_contracts/replay_event_stream_artifact_smoke_long_test.go` | Replay event-stream file conversion smoke owner | File-path opening, adjacent-factory hydration, and checked-in sample conversion helpers for the long replay smoke should stay test-owned here rather than expanding `pkg/replay` with production-facing file-wrapper convenience API. |
 | `tests/functional/runtime_api/api_generated_schema_deserialization_smoke_test.go` | Generated-schema runtime payload regression surface | Public workstation export or replay-contract retirements should keep serialized `factoryapi.Factory` assertions here so HTTP initial-structure, replay run-started, and runtime reconstruction lanes fail if emitted workstation JSON reintroduces retired keys like `promptTemplate`. |
 | `pkg/factory/runtime/dispatch_result_hook_test.go` | Dispatch-hook regression surface | Planner-present synchronous execution, planner-absent async worker-pool submission, missing-runner failures, and replay-tick validation should stay covered here when the behavior being changed is local to `workerPoolDispatchResultHook`, instead of forcing every dispatch-hook regression through the broader factory integration suite. |
 | `pkg/factory/subsystems/subsystem_dispatcher.go` | Dispatcher snapshot-enrichment owner | Dispatcher execution must reattach `d.state` onto raw runtime snapshots before handing them to scheduler/enablement flows, because runtime-only snapshots intentionally omit topology while runtime guards still need normalized transition-to-worker mappings. |
